# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API_Proj",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Employees/Get/{id}": {
      "get": {
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Employees/Update": {
      "put": {
        "tags": [
          "Employees"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Employees/Create": {
      "post": {
        "tags": [
          "Employees"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeForCreationDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeForCreationDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeForCreationDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Employees/Delete/{id}": {
      "delete": {
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Laptops/Get/{id}": {
      "get": {
        "tags": [
          "Laptops"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LaptopDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaptopDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaptopDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Laptops/Update": {
      "put": {
        "tags": [
          "Laptops"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LaptopDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LaptopDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LaptopDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Laptops/Update/{laptopID}/UnassignOwner": {
      "put": {
        "tags": [
          "Laptops"
        ],
        "parameters": [
          {
            "name": "laptopID",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Laptops/Create": {
      "post": {
        "tags": [
          "Laptops"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LaptopForCreationDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LaptopForCreationDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LaptopForCreationDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LaptopDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaptopDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LaptopDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Laptops/Delete/{id}": {
      "delete": {
        "tags": [
          "Laptops"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "EmployeeDTO": {
        "required": [
          "employeeID"
        ],
        "type": "object",
        "properties": {
          "employeeID": {
            "type": "integer",
            "format": "int32"
          },
          "employeeName": {
            "type": "string",
            "nullable": true
          },
          "jobTitle": {
            "type": "string",
            "nullable": true
          },
          "yearsAtCompany": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "currentProjects": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "officesIDs": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "laptopID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EmployeeForCreationDTO": {
        "required": [
          "employeeName"
        ],
        "type": "object",
        "properties": {
          "employeeName": {
            "minLength": 1,
            "type": "string"
          },
          "jobTitle": {
            "type": "string",
            "nullable": true
          },
          "yearsAtCompany": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "currentProjects": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "officesIDs": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "laptopID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LaptopDTO": {
        "required": [
          "laptopID"
        ],
        "type": "object",
        "properties": {
          "laptopID": {
            "type": "integer",
            "format": "int32"
          },
          "laptopName": {
            "type": "string",
            "nullable": true
          },
          "employeeID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LaptopForCreationDTO": {
        "required": [
          "laptopName"
        ],
        "type": "object",
        "properties": {
          "laptopName": {
            "minLength": 1,
            "type": "string"
          },
          "employeeID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}